Sync up with local changes to causes repo
Way back in ancient history we vendored a copy of the Facebooker
plug-in, which at the time advertised itself to be version 1.0.50 (it's
now at 1.0.75).

As time went by made a couple of changes to make things work nicely with
our application.

This commit replays those changes in our fork of the Facebooker repo, in
preparation for us submodulizing it.

Change-Id: I4e8c40590938b2770d792bb79df3effb454113f8
Reviewed-on: https://gerrit.causes.com/3501
Reviewed-by: Richard Zschech <richard.zschech@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>